requirements.txt
openai>=1.5.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.1